requirements.txt
altgraph==0.10.2
appnope==0.1.0
backports-abc==0.5
backports.functools-lru-cache==1.5
backports.shutil-get-terminal-size==1.0.0
bdist-mpkg==0.5.0
bonjour-py==0.3
cycler==0.10.0
decorator==4.3.0
enum34==1.1.6
futures==3.2.0
getch==1.0
ipykernel==4.10.0
ipython==5.8.0
ipython-genutils==0.2.0
jupyter-client==5.2.3
jupyter-core==4.4.0
kiwisolver==1.0.1
macholib==1.5.1
matplotlib==2.2.3
modulegraph==0.10.4
nose==1.3.7
numpy==1.15.2
pandas==0.23.4
pathlib2==2.3.2
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.3.0
prompt-toolkit==1.0.15
ptyprocess==0.6.0
py2app==0.7.3
Pygments==2.2.0
pyobjc-core==2.5.1
pyobjc-framework-Accounts==2.5.1
pyobjc-framework-AddressBook==2.5.1
pyobjc-framework-AppleScriptKit==2.5.1
pyobjc-framework-AppleScriptObjC==2.5.1
pyobjc-framework-Automator==2.5.1
pyobjc-framework-CFNetwork==2.5.1
pyobjc-framework-Cocoa==2.5.1
pyobjc-framework-Collaboration==2.5.1
pyobjc-framework-CoreData==2.5.1
pyobjc-framework-CoreLocation==2.5.1
pyobjc-framework-CoreText==2.5.1
pyobjc-framework-DictionaryServices==2.5.1
pyobjc-framework-EventKit==2.5.1
pyobjc-framework-ExceptionHandling==2.5.1
pyobjc-framework-FSEvents==2.5.1
pyobjc-framework-InputMethodKit==2.5.1
pyobjc-framework-InstallerPlugins==2.5.1
pyobjc-framework-InstantMessage==2.5.1
pyobjc-framework-LatentSemanticMapping==2.5.1
pyobjc-framework-LaunchServices==2.5.1
pyobjc-framework-Message==2.5.1
pyobjc-framework-OpenDirectory==2.5.1
pyobjc-framework-PreferencePanes==2.5.1
pyobjc-framework-PubSub==2.5.1
pyobjc-framework-QTKit==2.5.1
pyobjc-framework-Quartz==2.5.1
pyobjc-framework-ScreenSaver==2.5.1
pyobjc-framework-ScriptingBridge==2.5.1
pyobjc-framework-SearchKit==2.5.1
pyobjc-framework-ServiceManagement==2.5.1
pyobjc-framework-Social==2.5.1
pyobjc-framework-SyncServices==2.5.1
pyobjc-framework-SystemConfiguration==2.5.1
pyobjc-framework-WebKit==2.5.1
pyOpenSSL==0.13.1
pyparsing==2.2.2
python-dateutil==2.7.3
pytz==2018.5
pyzmq==17.1.2
qibuild==3.11.18
scandir==1.9.0
scipy==1.1.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
subprocess32==3.5.3
tornado==5.1.1
traitlets==4.3.2
wcwidth==0.1.7
xattr==0.6.4
zope.interface==4.1.1
